@@ -5,28 +5,29 @@
     "keywords": ["laravel", "framework"],
     "license": "MIT",
     "require": {
-        "php": "^8.2",
-        "bepsvpt/secure-headers": "^7.5",
-        "itsgoingd/clockwork": "^5.2",
-        "laravel/framework": "^11.9",
+        "php": "^8.3",
+        "bepsvpt/secure-headers": "^9.0",
+        "itsgoingd/clockwork": "^5.3",
+        "laravel/framework": "^12.8",
         "laravel/sanctum": "^4.0",
-        "laravel/slack-notification-channel": "^3.2",
-        "laravel/tinker": "^2.9",
-        "league/flysystem-aws-s3-v3": "^3.28",
-        "rollbar/rollbar-laravel": "^8.0",
-        "spatie/laravel-translatable": "^6.7",
+        "laravel/slack-notification-channel": "^3.5",
+        "laravel/tinker": "^2.10",
+        "league/flysystem-aws-s3-v3": "^3.29",
+        "rollbar/rollbar-laravel": "^8.1",
+        "spatie/laravel-translatable": "^6.11",
         "spatie/laravel-translation-loader": "^2.8"
     },
     "require-dev": {
-        "brianium/paratest": "^7.4",
-        "fakerphp/faker": "^1.23",
-        "larastan/larastan": "^2.9",
-        "laravel-lang/common": "^6.3",
+        "brianium/paratest": "^7.8",
+        "fakerphp/faker": "^1.24",
+        "larastan/larastan": "^3.3",
+        "laravel-lang/common": "^6.7",
+        "laravel/pail": "^1.2.2",
         "laravel/pint": "^1.13",
         "laravel/sail": "^1.26",
         "mockery/mockery": "^1.6",
         "nunomaduro/collision": "^8.0",
-        "phpunit/phpunit": "^11.0.1"
+        "phpunit/phpunit": "^12.1"
     },
     "autoload": {
         "psr-4": {
@@ -55,6 +56,10 @@
             "@php artisan key:generate --ansi",
             "@php -r \"file_exists('database/database.sqlite') || touch('database/database.sqlite');\"",
             "@php artisan migrate --graceful --ansi"
+        ],
+        "dev": [
+            "Composer\\Config::disableProcessTimeout",
+            "npx concurrently -c \"#93c5fd,#c4b5fd,#fb7185,#fdba74\" \"php artisan serve\" \"php artisan queue:listen --tries=1\" \"php artisan pail --timeout=0\" \"npm run dev\" --names=server,queue,logs,vite"
         ]
     },
     "extra": {
