@@ -2,8 +2,10 @@
 
 namespace App\Http\Controllers;
 
-use App\Exceptions\TranslatableException;
+use App\Exceptions\NormalizedException;
+use Closure;
 use Illuminate\Foundation\Bus\DispatchesJobs;
+use Illuminate\Http\Request;
 
 class UnauthenticatedController extends Controller
 {
@@ -11,13 +13,13 @@ class UnauthenticatedController extends Controller
 
     public function __construct()
     {
-        $this->middleware(function ($request, $next) {
+        $this->middleware(function (Request $request, Closure $next) {
             // Validate if there is no user authenticated
             if (!is_null($request->user())) {
-                throw new TranslatableException(
+                throw new NormalizedException(
                     500,
                     'Authenticated user in unauthenticated route.',
-                    'api.ERROR.SOMETHING_WENT_WRONG',
+                    __('api.ERROR.SOMETHING_WENT_WRONG'),
                 );
             }
 
