@@ -4,6 +4,6 @@ declare(strict_types=1);
 
 return [
     'failed'   => 'These credentials do not match our records.',
-    'password' => 'The password is incorrect.',
+    'password' => 'The provided password is incorrect.',
     'throttle' => 'Too many login attempts. Please try again in :seconds seconds.',
 ];
