@@ -3,99 +3,99 @@
 declare(strict_types=1);
 
 return [
-    'accepted'        => 'The :attribute must be accepted.',
-    'accepted_if'     => 'The :attribute must be accepted when :other is :value.',
-    'active_url'      => 'The :attribute is not a valid URL.',
-    'after'           => 'The :attribute must be a date after :date.',
-    'after_or_equal'  => 'The :attribute must be a date after or equal to :date.',
-    'alpha'           => 'The :attribute must only contain letters.',
-    'alpha_dash'      => 'The :attribute must only contain letters, numbers, dashes and underscores.',
-    'alpha_num'       => 'The :attribute must only contain letters and numbers.',
-    'array'           => 'The :attribute must be an array.',
+    'accepted'        => 'The :attribute field must be accepted.',
+    'accepted_if'     => 'The :attribute field must be accepted when :other is :value.',
+    'active_url'      => 'The :attribute field must be a valid URL.',
+    'after'           => 'The :attribute field must be a date after :date.',
+    'after_or_equal'  => 'The :attribute field must be a date after or equal to :date.',
+    'alpha'           => 'The :attribute field must only contain letters.',
+    'alpha_dash'      => 'The :attribute field must only contain letters, numbers, dashes, and underscores.',
+    'alpha_num'       => 'The :attribute field must only contain letters and numbers.',
+    'array'           => 'The :attribute field must be an array.',
     'ascii'           => 'The :attribute field must only contain single-byte alphanumeric characters and symbols.',
-    'before'          => 'The :attribute must be a date before :date.',
-    'before_or_equal' => 'The :attribute must be a date before or equal to :date.',
+    'before'          => 'The :attribute field must be a date before :date.',
+    'before_or_equal' => 'The :attribute field must be a date before or equal to :date.',
     'between'         => [
-        'array'   => 'The :attribute must have between :min and :max items.',
-        'file'    => 'The :attribute must be between :min and :max kilobytes.',
-        'numeric' => 'The :attribute must be between :min and :max.',
-        'string'  => 'The :attribute must be between :min and :max characters.',
+        'array'   => 'The :attribute field must have between :min and :max items.',
+        'file'    => 'The :attribute field must be between :min and :max kilobytes.',
+        'numeric' => 'The :attribute field must be between :min and :max.',
+        'string'  => 'The :attribute field must be between :min and :max characters.',
     ],
     'boolean'           => 'The :attribute field must be true or false.',
     'can'               => 'The :attribute field contains an unauthorized value.',
-    'confirmed'         => 'The :attribute confirmation does not match.',
+    'confirmed'         => 'The :attribute field confirmation does not match.',
     'contains'          => 'The :attribute field is missing a required value.',
     'current_password'  => 'The password is incorrect.',
-    'date'              => 'The :attribute is not a valid date.',
-    'date_equals'       => 'The :attribute must be a date equal to :date.',
-    'date_format'       => 'The :attribute does not match the format :format.',
+    'date'              => 'The :attribute field must be a valid date.',
+    'date_equals'       => 'The :attribute field must be a date equal to :date.',
+    'date_format'       => 'The :attribute field must match the format :format.',
     'decimal'           => 'The :attribute field must have :decimal decimal places.',
-    'declined'          => 'The :attribute must be declined.',
-    'declined_if'       => 'The :attribute must be declined when :other is :value.',
-    'different'         => 'The :attribute and :other must be different.',
-    'digits'            => 'The :attribute must be :digits digits.',
-    'digits_between'    => 'The :attribute must be between :min and :max digits.',
-    'dimensions'        => 'The :attribute has invalid image dimensions.',
+    'declined'          => 'The :attribute field must be declined.',
+    'declined_if'       => 'The :attribute field must be declined when :other is :value.',
+    'different'         => 'The :attribute field and :other must be different.',
+    'digits'            => 'The :attribute field must be :digits digits.',
+    'digits_between'    => 'The :attribute field must be between :min and :max digits.',
+    'dimensions'        => 'The :attribute field has invalid image dimensions.',
     'distinct'          => 'The :attribute field has a duplicate value.',
     'doesnt_end_with'   => 'The :attribute field must not end with one of the following: :values.',
     'doesnt_start_with' => 'The :attribute field must not start with one of the following: :values.',
-    'email'             => 'The :attribute must be a valid email address.',
-    'ends_with'         => 'The :attribute must end with one of the following: :values.',
-    'enum'              => 'The :attribute field value is not in the list of allowed values.',
-    'exists'            => 'The :attribute field value does not exist.',
+    'email'             => 'The :attribute field must be a valid email address.',
+    'ends_with'         => 'The :attribute field must end with one of the following: :values.',
+    'enum'              => 'The selected :attribute is invalid.',
+    'exists'            => 'The selected :attribute is invalid.',
     'extensions'        => 'The :attribute field must have one of the following extensions: :values.',
-    'file'              => 'The :attribute must be a file.',
+    'file'              => 'The :attribute field must be a file.',
     'filled'            => 'The :attribute field must have a value.',
     'gt'                => [
-        'array'   => 'The :attribute must have more than :value items.',
-        'file'    => 'The :attribute must be greater than :value kilobytes.',
-        'numeric' => 'The :attribute must be greater than :value.',
-        'string'  => 'The :attribute must be greater than :value characters.',
+        'array'   => 'The :attribute field must have more than :value items.',
+        'file'    => 'The :attribute field must be greater than :value kilobytes.',
+        'numeric' => 'The :attribute field must be greater than :value.',
+        'string'  => 'The :attribute field must be greater than :value characters.',
     ],
     'gte' => [
-        'array'   => 'The :attribute must have :value items or more.',
-        'file'    => 'The :attribute must be greater than or equal to :value kilobytes.',
-        'numeric' => 'The :attribute must be greater than or equal to :value.',
-        'string'  => 'The :attribute must be greater than or equal to :value characters.',
+        'array'   => 'The :attribute field must have :value items or more.',
+        'file'    => 'The :attribute field must be greater than or equal to :value kilobytes.',
+        'numeric' => 'The :attribute field must be greater than or equal to :value.',
+        'string'  => 'The :attribute field must be greater than or equal to :value characters.',
     ],
     'hex_color' => 'The :attribute field must be a valid hexadecimal color.',
-    'image'     => 'The :attribute must be an image.',
-    'in'        => 'The :attribute field value is not in the list of allowed values.',
-    'in_array'  => 'The :attribute field does not exist in :other.',
-    'integer'   => 'The :attribute must be an integer.',
-    'ip'        => 'The :attribute must be a valid IP address.',
-    'ipv4'      => 'The :attribute must be a valid IPv4 address.',
-    'ipv6'      => 'The :attribute must be a valid IPv6 address.',
-    'json'      => 'The :attribute must be a valid JSON string.',
+    'image'     => 'The :attribute field must be an image.',
+    'in'        => 'The selected :attribute is invalid.',
+    'in_array'  => 'The :attribute field must exist in :other.',
+    'integer'   => 'The :attribute field must be an integer.',
+    'ip'        => 'The :attribute field must be a valid IP address.',
+    'ipv4'      => 'The :attribute field must be a valid IPv4 address.',
+    'ipv6'      => 'The :attribute field must be a valid IPv6 address.',
+    'json'      => 'The :attribute field must be a valid JSON string.',
     'list'      => 'The :attribute field must be a list.',
     'lowercase' => 'The :attribute field must be lowercase.',
     'lt'        => [
-        'array'   => 'The :attribute must have less than :value items.',
-        'file'    => 'The :attribute must be less than :value kilobytes.',
-        'numeric' => 'The :attribute must be less than :value.',
-        'string'  => 'The :attribute must be less than :value characters.',
+        'array'   => 'The :attribute field must have less than :value items.',
+        'file'    => 'The :attribute field must be less than :value kilobytes.',
+        'numeric' => 'The :attribute field must be less than :value.',
+        'string'  => 'The :attribute field must be less than :value characters.',
     ],
     'lte' => [
-        'array'   => 'The :attribute must not have more than :value items.',
-        'file'    => 'The :attribute must be less than or equal to :value kilobytes.',
-        'numeric' => 'The :attribute must be less than or equal to :value.',
-        'string'  => 'The :attribute must be less than or equal to :value characters.',
+        'array'   => 'The :attribute field must not have more than :value items.',
+        'file'    => 'The :attribute field must be less than or equal to :value kilobytes.',
+        'numeric' => 'The :attribute field must be less than or equal to :value.',
+        'string'  => 'The :attribute field must be less than or equal to :value characters.',
     ],
-    'mac_address' => 'The :attribute must be a valid MAC address.',
+    'mac_address' => 'The :attribute field must be a valid MAC address.',
     'max'         => [
-        'array'   => 'The :attribute must not have more than :max items.',
-        'file'    => 'The :attribute must not be greater than :max kilobytes.',
-        'numeric' => 'The :attribute must not be greater than :max.',
-        'string'  => 'The :attribute must not be greater than :max characters.',
+        'array'   => 'The :attribute field must not have more than :max items.',
+        'file'    => 'The :attribute field must not be greater than :max kilobytes.',
+        'numeric' => 'The :attribute field must not be greater than :max.',
+        'string'  => 'The :attribute field must not be greater than :max characters.',
     ],
     'max_digits' => 'The :attribute field must not have more than :max digits.',
-    'mimes'      => 'The :attribute must be a file of type: :values.',
-    'mimetypes'  => 'The :attribute must be a file of type: :values.',
+    'mimes'      => 'The :attribute field must be a file of type: :values.',
+    'mimetypes'  => 'The :attribute field must be a file of type: :values.',
     'min'        => [
-        'array'   => 'The :attribute must have at least :min items.',
-        'file'    => 'The :attribute must be at least :min kilobytes.',
-        'numeric' => 'The :attribute must be at least :min.',
-        'string'  => 'The :attribute must be at least :min characters.',
+        'array'   => 'The :attribute field must have at least :min items.',
+        'file'    => 'The :attribute field must be at least :min kilobytes.',
+        'numeric' => 'The :attribute field must be at least :min.',
+        'string'  => 'The :attribute field must be at least :min characters.',
     ],
     'min_digits'       => 'The :attribute field must have at least :min digits.',
     'missing'          => 'The :attribute field must be missing.',
@@ -103,10 +103,10 @@ return [
     'missing_unless'   => 'The :attribute field must be missing unless :other is :value.',
     'missing_with'     => 'The :attribute field must be missing when :values is present.',
     'missing_with_all' => 'The :attribute field must be missing when :values are present.',
-    'multiple_of'      => 'The :attribute must be a multiple of :value.',
-    'not_in'           => 'The :attribute field must not be in the list.',
-    'not_regex'        => 'The :attribute format is invalid.',
-    'numeric'          => 'The :attribute must be a number.',
+    'multiple_of'      => 'The :attribute field must be a multiple of :value.',
+    'not_in'           => 'The selected :attribute is invalid.',
+    'not_regex'        => 'The :attribute field format is invalid.',
+    'numeric'          => 'The :attribute field must be a number.',
     'password'         => [
         'letters'       => 'The :attribute field must contain at least one letter.',
         'mixed'         => 'The :attribute field must contain at least one uppercase and one lowercase letter.',
@@ -114,42 +114,44 @@ return [
         'symbols'       => 'The :attribute field must contain at least one symbol.',
         'uncompromised' => 'The given :attribute has appeared in a data leak. Please choose a different :attribute.',
     ],
-    'present'              => 'The :attribute field must be present.',
-    'present_if'           => 'The :attribute field must be present when :other is :value.',
-    'present_unless'       => 'The :attribute field must be present unless :other is :value.',
-    'present_with'         => 'The :attribute field must be present when :values is present.',
-    'present_with_all'     => 'The :attribute field must be present when :values are present.',
-    'prohibited'           => 'The :attribute field is prohibited.',
-    'prohibited_if'        => 'The :attribute field is prohibited when :other is :value.',
-    'prohibited_unless'    => 'The :attribute field is prohibited unless :other is in :values.',
-    'prohibits'            => 'The :attribute field prohibits :other from being present.',
-    'regex'                => 'The :attribute format is invalid.',
-    'required'             => 'The :attribute field is required.',
-    'required_array_keys'  => 'The :attribute field must contain entries for: :values.',
-    'required_if'          => 'The :attribute field is required when :other is :value.',
-    'required_if_accepted' => 'The :attribute field is required when :other is accepted.',
-    'required_if_declined' => 'The :attribute field is required when :other is declined.',
-    'required_unless'      => 'The :attribute field is required unless :other is in :values.',
-    'required_with'        => 'The :attribute field is required when :values is present.',
-    'required_with_all'    => 'The :attribute field is required when :values are present.',
-    'required_without'     => 'The :attribute field is required when :values is not present.',
-    'required_without_all' => 'The :attribute field is required when none of :values are present.',
-    'same'                 => 'The :attribute and :other must match.',
-    'size'                 => [
-        'array'   => 'The :attribute must contain :size items.',
-        'file'    => 'The :attribute must be :size kilobytes.',
-        'numeric' => 'The :attribute must be :size.',
-        'string'  => 'The :attribute must be :size characters.',
+    'present'                => 'The :attribute field must be present.',
+    'present_if'             => 'The :attribute field must be present when :other is :value.',
+    'present_unless'         => 'The :attribute field must be present unless :other is :value.',
+    'present_with'           => 'The :attribute field must be present when :values is present.',
+    'present_with_all'       => 'The :attribute field must be present when :values are present.',
+    'prohibited'             => 'The :attribute field is prohibited.',
+    'prohibited_if'          => 'The :attribute field is prohibited when :other is :value.',
+    'prohibited_if_accepted' => 'The :attribute field is prohibited when :other is accepted.',
+    'prohibited_if_declined' => 'The :attribute field is prohibited when :other is declined.',
+    'prohibited_unless'      => 'The :attribute field is prohibited unless :other is in :values.',
+    'prohibits'              => 'The :attribute field prohibits :other from being present.',
+    'regex'                  => 'The :attribute field format is invalid.',
+    'required'               => 'The :attribute field is required.',
+    'required_array_keys'    => 'The :attribute field must contain entries for: :values.',
+    'required_if'            => 'The :attribute field is required when :other is :value.',
+    'required_if_accepted'   => 'The :attribute field is required when :other is accepted.',
+    'required_if_declined'   => 'The :attribute field is required when :other is declined.',
+    'required_unless'        => 'The :attribute field is required unless :other is in :values.',
+    'required_with'          => 'The :attribute field is required when :values is present.',
+    'required_with_all'      => 'The :attribute field is required when :values are present.',
+    'required_without'       => 'The :attribute field is required when :values is not present.',
+    'required_without_all'   => 'The :attribute field is required when none of :values are present.',
+    'same'                   => 'The :attribute field must match :other.',
+    'size'                   => [
+        'array'   => 'The :attribute field must contain :size items.',
+        'file'    => 'The :attribute field must be :size kilobytes.',
+        'numeric' => 'The :attribute field must be :size.',
+        'string'  => 'The :attribute field must be :size characters.',
     ],
-    'starts_with' => 'The :attribute must start with one of the following: :values.',
-    'string'      => 'The :attribute must be a string.',
-    'timezone'    => 'The :attribute must be a valid timezone.',
+    'starts_with' => 'The :attribute field must start with one of the following: :values.',
+    'string'      => 'The :attribute field must be a string.',
+    'timezone'    => 'The :attribute field must be a valid timezone.',
     'ulid'        => 'The :attribute field must be a valid ULID.',
     'unique'      => 'The :attribute has already been taken.',
     'uploaded'    => 'The :attribute failed to upload.',
     'uppercase'   => 'The :attribute field must be uppercase.',
-    'url'         => 'The :attribute must be a valid URL.',
-    'uuid'        => 'The :attribute must be a valid UUID.',
+    'url'         => 'The :attribute field must be a valid URL.',
+    'uuid'        => 'The :attribute field must be a valid UUID.',
     'attributes'  => [
         'address'                  => 'address',
         'affiliate_url'            => 'affiliate URL',
@@ -158,14 +160,17 @@ return [
         'announcement'             => 'announcement',
         'area'                     => 'area',
         'audience_prize'           => 'audience prize',
+        'audience_winner'          => 'audience winner',
         'available'                => 'available',
         'birthday'                 => 'birthday',
         'body'                     => 'body',
         'city'                     => 'city',
+        'company'                  => 'company',
         'compilation'              => 'compilation',
         'concept'                  => 'concept',
         'conditions'               => 'conditions',
         'content'                  => 'content',
+        'contest'                  => 'contest',
         'country'                  => 'country',
         'cover'                    => 'cover',
         'created_at'               => 'created at',
@@ -270,6 +275,8 @@ return [
         'user'                     => 'user',
         'username'                 => 'username',
         'value'                    => 'value',
+        'winner'                   => 'winner',
+        'work'                     => 'work',
         'year'                     => 'year',
     ],
 ];
