@@ -24,7 +24,7 @@ return [
     'boolean'           => 'O campo :attribute deve conter o valor verdadeiro ou falso.',
     'can'               => 'O campo :attribute contém um valor não autorizado.',
     'confirmed'         => 'A confirmação para o campo :attribute não coincide.',
-    'contains'          => 'The :attribute field is missing a required value.',
+    'contains'          => 'O campo :attribute não contém um valor obrigatório.',
     'current_password'  => 'A senha está incorreta.',
     'date'              => 'O campo :attribute não contém uma data válida.',
     'date_equals'       => 'O campo :attribute deve ser uma data igual a :date.',
@@ -114,28 +114,30 @@ return [
         'symbols'       => 'O :attribute deve conter pelo menos um símbolo.',
         'uncompromised' => 'O dado :attribute apareceu em um vazamento de dados. Por favor, escolha um :attribute diferente.',
     ],
-    'present'              => 'O campo :attribute deve estar presente.',
-    'present_if'           => 'O campo :attribute deve estar presente quando :other for :value.',
-    'present_unless'       => 'O campo :attribute deve estar presente, a menos que :other seja :value.',
-    'present_with'         => 'O campo :attribute deve estar presente quando :values estiver presente.',
-    'present_with_all'     => 'O campo :attribute deve estar presente quando :values estiverem presentes.',
-    'prohibited'           => 'O campo :attribute é proibido.',
-    'prohibited_if'        => 'O campo :attribute é proibido quando :other é :value.',
-    'prohibited_unless'    => 'O campo :attribute é proibido a menos que :other esteja em :values.',
-    'prohibits'            => 'O campo :attribute proíbe :other de estar presente.',
-    'regex'                => 'O formato do valor informado no campo :attribute é inválido.',
-    'required'             => 'O campo :attribute é obrigatório.',
-    'required_array_keys'  => 'O campo :attribute deve conter entradas para: :values',
-    'required_if'          => 'O campo :attribute é obrigatório quando o valor do campo :other é igual a :value.',
-    'required_if_accepted' => 'O campo :attribute é obrigatório quando :other é aceito.',
-    'required_if_declined' => 'The :attribute field is required when :other is declined.',
-    'required_unless'      => 'O campo :attribute é obrigatório a menos que :other esteja presente em :values.',
-    'required_with'        => 'O campo :attribute é obrigatório quando :values está presente.',
-    'required_with_all'    => 'O campo :attribute é obrigatório quando um dos :values está presente.',
-    'required_without'     => 'O campo :attribute é obrigatório quando :values não está presente.',
-    'required_without_all' => 'O campo :attribute é obrigatório quando nenhum dos :values está presente.',
-    'same'                 => 'Os campos :attribute e :other devem conter valores iguais.',
-    'size'                 => [
+    'present'                => 'O campo :attribute deve estar presente.',
+    'present_if'             => 'O campo :attribute deve estar presente quando :other for :value.',
+    'present_unless'         => 'O campo :attribute deve estar presente, a menos que :other seja :value.',
+    'present_with'           => 'O campo :attribute deve estar presente quando :values estiver presente.',
+    'present_with_all'       => 'O campo :attribute deve estar presente quando :values estiverem presentes.',
+    'prohibited'             => 'O campo :attribute é proibido.',
+    'prohibited_if'          => 'O campo :attribute é proibido quando :other é :value.',
+    'prohibited_if_accepted' => 'O campo :attribute é proibido quando :other é aceito.',
+    'prohibited_if_declined' => 'O campo :attribute é proibido quando :other é recusado.',
+    'prohibited_unless'      => 'O campo :attribute é proibido a menos que :other esteja em :values.',
+    'prohibits'              => 'O campo :attribute proíbe :other de estar presente.',
+    'regex'                  => 'O formato do valor informado no campo :attribute é inválido.',
+    'required'               => 'O campo :attribute é obrigatório.',
+    'required_array_keys'    => 'O campo :attribute deve conter entradas para: :values',
+    'required_if'            => 'O campo :attribute é obrigatório quando o valor do campo :other é igual a :value.',
+    'required_if_accepted'   => 'O campo :attribute é obrigatório quando :other é aceito.',
+    'required_if_declined'   => 'O campo :attribute é obrigatório quando :other não for informado.',
+    'required_unless'        => 'O campo :attribute é obrigatório a menos que :other esteja presente em :values.',
+    'required_with'          => 'O campo :attribute é obrigatório quando :values está presente.',
+    'required_with_all'      => 'O campo :attribute é obrigatório quando um dos :values está presente.',
+    'required_without'       => 'O campo :attribute é obrigatório quando :values não está presente.',
+    'required_without_all'   => 'O campo :attribute é obrigatório quando nenhum dos :values está presente.',
+    'same'                   => 'Os campos :attribute e :other devem conter valores iguais.',
+    'size'                   => [
         'array'   => 'O campo :attribute deve conter :size itens.',
         'file'    => 'O campo :attribute deve conter um arquivo com o tamanho de :size kilobytes.',
         'numeric' => 'O campo :attribute deve conter o número :size.',
@@ -158,14 +160,17 @@ return [
         'announcement'             => 'anúncio',
         'area'                     => 'área',
         'audience_prize'           => 'prêmio do público',
+        'audience_winner'          => 'vencedor do público',
         'available'                => 'disponível',
         'birthday'                 => 'aniversário',
         'body'                     => 'conteúdo',
         'city'                     => 'cidade',
+        'company'                  => 'empresa',
         'compilation'              => 'compilação',
         'concept'                  => 'conceito',
         'conditions'               => 'condições',
         'content'                  => 'conteúdo',
+        'contest'                  => 'contestar',
         'country'                  => 'país',
         'cover'                    => 'cobrir',
         'created_at'               => 'criado em',
@@ -251,7 +256,7 @@ return [
         'specialization'           => 'especialização',
         'started_at'               => 'começou às',
         'state'                    => 'estado',
-        'status'                   => 'status',
+        'status'                   => 'estado',
         'street'                   => 'rua',
         'student'                  => 'estudante',
         'subject'                  => 'assunto',
@@ -270,6 +275,8 @@ return [
         'user'                     => 'do utilizador',
         'username'                 => 'usuário',
         'value'                    => 'valor',
+        'winner'                   => 'ganhador',
+        'work'                     => 'trabalho',
         'year'                     => 'ano',
     ],
 ];
